[feat] #83 settlement testcode 정리
